requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
jinja2>=3.1.0
python-multipart>=0.0.18
pyyaml>=6.0.2
python-dotenv>=1.0.1
